Refactor: 절대경로로 수정
What is this PR?
- 절대경로로 수정

Todo
