+++ client/src/components/card/feed-card/FeedCard.tsx
@@ -6,8 +6,7 @@ import { useReadLocalStorage } from 'usehooks-ts';
 import 'swiper/css';
 import 'swiper/css/pagination';
 import 'swiper/css/navigation';
-import Profile from '../../../common/profile/Profile.tsx';
-import '../../../common/carousel/carousel.css';
+import Profile from '@/common/profile/Profile.tsx';
 import {
   Container,
   Feed,
@@ -16,7 +15,8 @@ import {
   UserName,
   Context,
   More,
-} from './FeedCard.styled.tsx';
+} from '@/components/card/feed-card/FeedCard.styled.tsx';
+import '@/common/carousel/carousel.css';
 
 interface ImagesStructure {
   imageId: number;
