+++ client/src/components/card/walk-card/walkCard.tsx
@@ -1,8 +1,3 @@
-import { ReactComponent as Paw } from '../../../assets/button/like.svg';
-import { ReactComponent as Dog } from '../../../assets/dog.svg';
-import { ReactComponent as Pin } from '../../../assets/pin.svg';
-import Profile from '../../../common/profile/Profile.tsx';
-import { BooleanStr } from '../../../types/propType.ts';
 import {
   Backdrop,
   CardContainer,
@@ -15,6 +10,11 @@ import {
   WriterProfile,
   CompleteText,
 } from './walkCard.styled';
+import { ReactComponent as Paw } from '@/assets/button/like.svg';
+import { ReactComponent as Dog } from '@/assets/dog.svg';
+import { ReactComponent as Pin } from '@/assets/pin.svg';
+import Profile from '@/common/profile/Profile.tsx';
+import { BooleanStr } from '@/types/propType.ts';
 
 type walkCardProps = {
   time: string;
