+++ client/src/components/notice/NoticeNoData.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
-import LyingDownDog from '../../assets/illustration/lying-down-dog.png';
-import { ErrorText, HomeBtn } from '../../pages/notFound/NotFound.styled.tsx';
+import LyingDownDog from '@/assets/illustration/lying-down-dog.png';
+import { ErrorText, HomeBtn } from '@/pages/notFound/NotFound.styled.tsx';
 
 type NoticeNoDataProp = {
   url: string;
