+++ client/src/components/notice/NoticeNotWrite.tsx
@@ -1,5 +1,5 @@
-import Cat404 from '../../assets/illustration/404cat.png';
-import { ErrorText } from '../../pages/notFound/NotFound.styled.tsx';
+import Cat404 from '@/assets/illustration/404cat.png';
+import { ErrorText } from '@/pages/notFound/NotFound.styled.tsx';
 
 export default function NoticeNotWrite() {
   return (
