+++ client/src/components/notice/NoticeOnlyOwner.tsx
@@ -1,6 +1,6 @@
 import { useMatch, useNavigate } from 'react-router-dom';
-import LoginPets from '../../assets/illustration/loginpet.png';
-import { ErrorText, HomeBtn } from '../../pages/notFound/NotFound.styled.tsx';
+import LoginPets from '@/assets/illustration/loginpet.png';
+import { ErrorText, HomeBtn } from '@/pages/notFound/NotFound.styled.tsx';
 
 export default function NoticeOnlyOwner() {
   const navigate = useNavigate();
