+++ client/src/components/notice/NoticeServerError.tsx
@@ -1,6 +1,6 @@
 import { useNavigate } from 'react-router-dom';
-import Cat404 from '../../assets/illustration/404cat.png';
-import { ErrorText, HomeBtn } from '../../pages/notFound/NotFound.styled.tsx';
+import Cat404 from '@/assets/illustration/404cat.png';
+import { ErrorText, HomeBtn } from '@/pages/notFound/NotFound.styled.tsx';
 
 export default function NoticeServerError() {
   const navigate = useNavigate();
