--- client/src/pages/copyPopUp/CopyPopUp.tsx
@@ -1,5 +1,0 @@
-export function Component() {
-  return <div>CopyPopUp</div>;
-}
-
-Component.displayName = 'CopyPopUp';
