+++ client/src/pages/home/Home.tsx
@@ -9,21 +9,21 @@ import 'swiper/css';
 import 'swiper/css/pagination';
 import 'swiper/css/navigation';
 import { useLocalStorage, useReadLocalStorage } from 'usehooks-ts';
-import { getGuestFeedList, getHostFeedList } from '../../api/queryfn.ts';
-import { SERVER_URL } from '../../api/url.ts';
-import LyingDownDog from '../../assets/illustration/lying-down-dog.png';
-import PetFriends from '../../assets/illustration/pet-friends.png';
-import { PopupBackGround } from '../../common/popup/popup.styled.tsx';
-import Popup from '../../common/popup/Popup.tsx';
-import FeedCard from '../../components/card/feed-card/FeedCard.tsx';
-import SideNav from '../../components/card/sidenav/SideNav.tsx';
-import LoadingComponent from '../../components/loading/LoadingComponent.tsx';
-import NoticeServerError from '../../components/notice/NoticeServerError.tsx';
-import Toast from '../../components/toast/Toast.tsx';
-import { Feed } from '../../types/feedTypes.ts';
-import CircleProgressBar from './CricleProgressBar.tsx';
-import { Container, ReBtn, TopBtn } from './Home.styled.tsx';
-import '../../common/carousel/carousel.css';
+import { getGuestFeedList, getHostFeedList } from '@/api/queryfn.ts';
+import { SERVER_URL } from '@/api/url.ts';
+import LyingDownDog from '@/assets/illustration/lying-down-dog.png';
+import PetFriends from '@/assets/illustration/pet-friends.png';
+import { PopupBackGround } from '@/common/popup/popup.styled.tsx';
+import Popup from '@/common/popup/Popup.tsx';
+import FeedCard from '@/components/card/feed-card/FeedCard.tsx';
+import SideNav from '@/components/card/sidenav/SideNav.tsx';
+import LoadingComponent from '@/components/loading/LoadingComponent.tsx';
+import NoticeServerError from '@/components/notice/NoticeServerError.tsx';
+import Toast from '@/components/toast/Toast.tsx';
+import CircleProgressBar from '@/pages/home/CricleProgressBar.tsx';
+import { Container, ReBtn, TopBtn } from '@/pages/home/Home.styled.tsx';
+import { Feed } from '@/types/feedTypes.ts';
+import '@/common/carousel/carousel.css';
 
 export function Component() {
   const queryClient = useQueryClient();
