+++ client/vite.config.ts
@@ -15,4 +15,7 @@ export default defineConfig({
       },
     },
   },
+  resolve: {
+    alias: [{ find: '@', replacement: '/src' }],
+  },
 });
